refactor: split pipeline editor components
