+++ frontend/src/lib/components/PipelineEditor.svelte
@@ -1,43 +1,9 @@
 <script lang="ts">
   import Button from './Button.svelte';
-  import { onMount, createEventDispatcher } from 'svelte'; // Import onMount and createEventDispatcher
+  import StageList from './pipeline_editor/StageList.svelte';
+  import { onMount, createEventDispatcher } from 'svelte';
   import { apiFetch } from '$lib/utils/apiUtils';
-
-  // Define Stage interface for clarity and new fields
-  export interface Stage { // Exporting if it's useful for other components, otherwise keep local
-    id: string; // Client-side unique ID
-    type: string;
-    command?: string | null;
-    prompt_name?: string | null;
-
-    // New fields for OCR stage configuration
-    ocr_engine?: 'default' | 'external';
-    ocr_stage_endpoint?: string | null;
-    ocr_stage_key?: string | null;
-
-    config?: {
-      strategy?: string;
-      parameters?: any;
-      template?: string;
-      summaryFields?: string[];
-      _summaryFieldsString?: string;
-    } | null;
-  }
-
-  interface RegexPatternConfig {
-    id: string;
-    name: string;
-    regex: string;
-    captureGroupIndex?: number;
-  }
-
-  // Define Pipeline structure, matching +layout.svelte and backend expectations
-  export interface Pipeline {
-    id?: string;
-    name: string;
-    org_id: string;
-    stages: Stage[];
-  }
+  import type { Stage, Pipeline, EditorPromptTemplate, RegexPatternConfig } from './pipeline_editor/types';
 
   export let orgId: string;
   export let initialPipeline: Pipeline | null = null;
@@ -55,10 +21,6 @@
   }
 
   // State for available prompt templates
-  interface EditorPromptTemplate {
-    name: string;
-    text: string;
-  }
   let availablePromptTemplates: EditorPromptTemplate[] = [];
   let isLoadingOrgSettings: boolean = false;
   let promptTemplatesError: string | null = null;
@@ -187,49 +149,6 @@
     }
   }
 
-  function initializeParseStrategyParameters(stage: Stage) {
-    if (stage.type.toLowerCase() !== 'parse') {
-        // This function should only be called for parse stages, but as a safeguard:
-        if (stage.config) stage.config = undefined; // Clear config if type changed away from parse
-        pipeline.stages = [...pipeline.stages];
-        return;
-    }
-
-    // Ensure config object exists for a parse stage
-    if (!stage.config) {
-        stage.config = { strategy: 'Passthrough', parameters: {} };
-    }
-    // stage.config is now guaranteed to exist
-
-    switch (stage.config.strategy) {
-        case 'KeywordExtraction':
-            stage.config.parameters = { keywords: [''], caseSensitive: false };
-            break;
-        case 'RegexExtraction':
-            stage.config.parameters = { patterns: [{ id: Date.now().toString() + Math.random().toString(36).substring(2,9), name: '', regex: '' }] };
-            break;
-        case 'SimpleTableExtraction':
-            stage.config.parameters = { _headerKeywordsString: '', _stopKeywordsString: '', headerKeywords: [], stopKeywords: [] };
-            break;
-        case 'RegexExtraction': // Ensure this also initializes with captureGroupIndex
-            stage.config.parameters = {
-                patterns: [{
-                  id: Date.now().toString() + Math.random().toString(36).substring(2,9),
-                  name: '',
-                  regex: '',
-                  captureGroupIndex: 1 // Default for new patterns in UI
-                }]
-            };
-            break;
-        case 'Passthrough':
-        default:
-            stage.config.parameters = {};
-            break;
-    }
-    // Force reactivity for the entire stages array to ensure UI updates for nested changes
-    // Force reactivity for the entire stages array to ensure UI updates for nested changes
-    pipeline.stages = [...pipeline.stages];
-  }
 
   // This function prepares the pipeline data for saving (e.g., removing temporary UI fields)
   function getSanitizedPipelineForSave(): Pipeline {
@@ -374,100 +293,9 @@
   }
   */
 
-  function removeStage(idToRemove: string) {
-    pipeline.stages = pipeline.stages.filter(stage => stage.id !== idToRemove);
-  }
-
-  // State for drag and drop (using itemId based logic)
-  let draggedItemId: string | null = null;
-  let draggedOverIndex: number | null = null; // Visual index for hover effect
-  // draggingIndex (numeric index of dragged item) is not strictly needed if using ID, but can be kept for visuals if preferred
-  let draggingVisualIndex: number | null = null;
-
-
-  function handleDragStart(event: DragEvent, stageId: string, index: number) {
-    draggedItemId = stageId;
-    draggingVisualIndex = index; // For visual feedback like 'dragging' class
-    event.dataTransfer!.setData('text/plain', stageId);
-    event.dataTransfer!.effectAllowed = 'move';
-  }
-
-  function handleDragOver(event: DragEvent, targetIndex: number) {
-    event.preventDefault();
-    if (draggedItemId === null) return;
-
-    const targetStageId = pipeline.stages[targetIndex]?.id;
-    if (draggedItemId !== targetStageId) { // Don't highlight if dragging over itself based on ID
-      draggedOverIndex = targetIndex;
-    } else {
-      draggedOverIndex = null;
-    }
-    event.dataTransfer!.dropEffect = 'move';
-  }
-
-  function handleDragEnter(event: DragEvent, targetIndex: number) {
-    event.preventDefault();
-    if (draggedItemId === null) return;
-    const targetStageId = pipeline.stages[targetIndex]?.id;
-    if (draggedItemId !== targetStageId) {
-      draggedOverIndex = targetIndex;
-    }
-  }
-
-  function handleDragLeave(event: DragEvent) {
-    const currentTarget = event.currentTarget as HTMLElement;
-    const relatedTarget = event.relatedTarget as HTMLElement | null;
-    if (!relatedTarget || !currentTarget.contains(relatedTarget)) {
-      // Cleared by dragover on new item or dragend
-    }
-  }
-
-  function handleDrop(event: DragEvent, targetItemIndex: number) {
-    event.preventDefault();
-    if (!draggedItemId) return;
-
-    const draggedItemOriginalIndex = pipeline.stages.findIndex(s => s.id === draggedItemId);
-    if (draggedItemOriginalIndex === -1) return; // Should not happen
-
-    // If dropped on itself effectively (same index after potential shifts)
-    // This logic can be tricky if targetItemIndex is the final visual index.
-    // It's often simpler to just re-construct the array based on splice/insert.
-
-    const items = [...pipeline.stages];
-    const [draggedItem] = items.splice(draggedItemOriginalIndex, 1); // Get the dragged item
-
-    // Insert at the target drop index
-    items.splice(targetItemIndex, 0, draggedItem);
-
-    pipeline.stages = items;
-
-    draggedItemId = null;
-    draggingIndex = null; // Reset visual index if used
-    draggedOverIndex = null;
-  }
-
-  function handleDragEnd() {
-    draggedItemId = null;
-    draggingVisualIndex = null; // Reset visual index
-    draggedOverIndex = null;
-  }
 </script>
 
 <style>
-  .stage-item {
-    transition: background-color 0.2s ease-out, border 0.2s ease-out;
-    border: 2px solid transparent; /* Reserve space for border highlight */
-  }
-  .dragging {
-    opacity: 0.5;
-    background-color: rgba(255, 255,255, 0.1); /* Slight background change */
-  }
-  .drag-over-highlight {
-    /* Visual cue for where the item will be dropped */
-    /* E.g., a border on the item it will be inserted before/after */
-    border-color: #3b82f6; /* Tailwind blue-500 */
-    background-color: rgba(59, 130, 246, 0.1); /* Light blue background */
-  }
 </style>
 
 <div class="space-y-4 text-gray-200">
@@ -477,237 +305,9 @@
       {promptTemplatesError}
     </div>
   {/if}
-  <div class="space-y-3">
-    {#each pipeline.stages as stage, i (stage.id)}
-      <div
-        class={`stage-item p-4 rounded-lg cursor-grab border-2 ${
-          draggingVisualIndex === i ? 'dragging !border-accent' : 'border-neutral-700/70'
-        } ${
-          draggedOverIndex === i && draggedItemId !== stage.id
-            ? 'drag-over-highlight !border-accent'
-            : 'hover:border-neutral-600'
-        }`}
-        draggable="true"
-        on:dragstart={(event) => handleDragStart(event, stage.id, i)}
-        on:dragover={(event) => handleDragOver(event, i)}
-        on:dragenter={(event) => handleDragEnter(event, i)}
-        on:dragleave={handleDragLeave}
-        on:drop={(event) => handleDrop(event, i)}
-        on:dragend={handleDragEnd}
-      >
-        <div class="flex items-center gap-2 mb-2">
-          <input class="glass-input flex-1 !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100" bind:value={stage.type} placeholder="Stage Type (e.g., ocr, ai)"/>
-          <input class="glass-input flex-1 !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100" bind:value={stage.command} placeholder="Command / Config (optional)" />
-          <Button variant="ghost" customClass="!text-red-500 hover:!text-red-400 !p-1" on:click={() => removeStage(stage.id)}>
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" class="w-5 h-5"><path fill-rule="evenodd" d="M8.75 1A2.75 2.75 0 006 3.75v.443c-.795.077-1.58.176-2.365.298a.75.75 0 10.23 1.482l.149-.022.841 10.518A2.75 2.75 0 007.596 19h4.807a2.75 2.75 0 002.742-2.53l.841-10.52.149.023a.75.75 0 00.23-1.482A41.03 41.03 0 0014 4.193V3.75A2.75 2.75 0 0011.25 1h-2.5zM10 4c.84 0 1.673.025 2.5.075V3.75c0-.69-.56-1.25-1.25-1.25h-2.5c-.69 0-1.25.56-1.25 1.25V4c.827-.05 1.66-.075 2.5-.075zM8.088 15.816a1.25 1.25 0 01-1.24-1.22L6.003 6.51a.75.75 0 111.494-.138l.84 8.088a1.25 1.25 0 01-1.24 1.221zM11.912 15.816a1.25 1.25 0 01-1.24-1.221l-.84-8.088a.75.75 0 111.494.138l.84 8.088a1.25 1.25 0 01-1.24 1.22z" clip-rule="evenodd" /></svg>
-          </Button>
-        </div>
-        {#if stage.type === 'ai'}
-          <div class="form-group pt-2 border-t border-neutral-700/50">
-            <label for={`stage-prompt-${stage.id}`} class="block text-xs font-light text-gray-300 mb-1">
-              Prompt Template
-            </label>
-            {#if isLoadingOrgSettings}
-              <p class="text-sm font-light text-gray-400 dark:text-gray-500 py-2">Loading prompt templates...</p>
-            {:else if availablePromptTemplates.length > 0}
-              <select
-                id={`stage-prompt-${stage.id}`}
-                bind:value={stage.prompt_name}
-                class="glass-input w-full text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-              >
-                <option value={undefined}>Default (No specific template)</option>
-                {#each availablePromptTemplates as template (template.name)}
-                  <option value={template.name}>{template.name}</option>
-                {/each}
-              </select>
-            {:else}
-              <p class="text-sm font-light text-gray-400 dark:text-gray-500 py-2">No prompt templates defined. AI uses default behavior.</p>
-            {/if}
-            {#if stage.prompt_name && availablePromptTemplates.find(p => p.name === stage.prompt_name)}
-                {@const selectedTemplateText = availablePromptTemplates.find(p => p.name === stage.prompt_name)?.text}
-                {#if selectedTemplateText}
-                    <div class="mt-1.5 p-1.5 bg-black/30 rounded text-xs text-gray-400 max-h-24 overflow-y-auto custom-scrollbar border border-neutral-600/50">
-                        <strong class="text-gray-300">Preview:</strong>
-                        <pre class="whitespace-pre-wrap font-mono text-[0.7rem] leading-snug">{selectedTemplateText.substring(0, 150)}{selectedTemplateText.length > 150 ? '...' : ''}</pre>
-                    </div>
-                {/if}
-            {/if}
-          </div>
-        {/if}
-        {#if stage.type.toLowerCase() === 'ocr'}
-          <div class="form-group mt-2 pt-2 border-t border-neutral-700/50 space-y-2">
-            <div>
-              <label for={`stage-ocr-engine-${stage.id}`} class="block text-xs font-light text-gray-300 mb-1">OCR Engine</label>
-              <select bind:value={stage.ocr_engine} id={`stage-ocr-engine-${stage.id}`} class="glass-input w-full text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100">
-                <option value="default">Default (Tesseract/Local)</option>
-                <option value="external">External API</option>
-              </select>
-            </div>
-
-            {#if stage.ocr_engine === 'external'}
-              <div class="mt-2 space-y-2 pl-2 border-l-2 border-neutral-700/40 ml-1">
-                <div class="pt-1">
-                  <label for={`stage-ocr-endpoint-${stage.id}`} class="block text-xs font-light text-gray-300 mb-1">
-                    Stage OCR API Endpoint
-                  </label>
-                  <input
-                    type="text"
-                    id={`stage-ocr-endpoint-${stage.id}`}
-                    bind:value={stage.ocr_stage_endpoint}
-                    class="glass-input w-full text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-                    placeholder="Overrides global OCR endpoint"
-                  />
-                </div>
-                <div>
-                  <label for={`stage-ocr-key-${stage.id}`} class="block text-xs font-light text-gray-300 mb-1">
-                    Stage OCR API Key
-                  </label>
-                  <input
-                    type="password"
-                    id={`stage-ocr-key-${stage.id}`}
-                    bind:value={stage.ocr_stage_key}
-                    class="glass-input w-full text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-                    placeholder="Overrides global OCR key"
-                  />
-                </div>
-              </div>
-            {/if}
-          </div>
-        {/if}
-
-        <!-- Parse Stage Configuration -->
-        {#if stage.type.toLowerCase() === 'parse'}
-          <div class="form-group mt-3 pt-3 border-t border-neutral-700/50 space-y-3">
-            <div>
-              <label for={`stage-parse-strategy-${stage.id}`} class="block text-xs font-medium text-gray-300 mb-1">
-                Parsing Strategy
-              </label>
-              <select
-                id={`stage-parse-strategy-${stage.id}`}
-                bind:value={stage.config.strategy}
-                on:change={() => initializeParseStrategyParameters(stage)}
-                class="glass-input w-full !text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-              >
-                <option value="Passthrough">Passthrough (Basic Lines/Text)</option>
-                <option value="KeywordExtraction">Keyword Extraction</option>
-                <option value="RegexExtraction">Regex Extraction</option>
-                <option value="SimpleTableExtraction">Simple Table Extraction (Basic Placeholder)</option>
-              </select>
-            </div>
-
-            <!-- Parameters for KeywordExtraction -->
-            {#if stage.config?.strategy === 'KeywordExtraction'}
-              <div class="pl-3 border-l-2 border-neutral-700 space-y-2 py-2 text-xs">
-                <label class="block font-medium text-gray-300 mb-1">Keywords:</label>
-                {#each stage.config.parameters.keywords as keyword, k (k)}
-                  <div class="flex items-center space-x-2">
-                    <input type="text" bind:value={stage.config.parameters.keywords[k]} class="glass-input flex-grow !text-xs !bg-neutral-500/40" placeholder="Enter keyword"/>
-                    <Button variant="ghost" customClass="!px-1.5 !py-0.5 !text-red-400 hover:!text-red-300" on:click={() => stage.config.parameters.keywords = stage.config.parameters.keywords.filter((_, idx) => idx !== k)}>X</Button>
-                  </div>
-                {/each}
-                <Button variant="secondary" customClass="!text-xs !py-1" on:click={() => stage.config.parameters.keywords = [...stage.config.parameters.keywords, '']}>Add Keyword</Button>
-                <label class="flex items-center space-x-2 mt-2 cursor-pointer">
-                  <input type="checkbox" bind:checked={stage.config.parameters.caseSensitive} class="form-checkbox h-4 w-4 text-accent rounded !bg-neutral-700 border-neutral-600 focus:ring-accent/50"/>
-                  <span class="text-gray-300">Case Sensitive</span>
-                </label>
-              </div>
-            {/if}
-
-            <!-- Parameters for RegexExtraction -->
-            {#if stage.config?.strategy === 'RegexExtraction'}
-              <div class="pl-3 border-l-2 border-neutral-700 space-y-2 py-2 text-xs">
-                <label class="block font-medium text-gray-300 mb-1">Regex Patterns:</label>
-                {#each stage.config.parameters.patterns as pattern, k (pattern.id)}
-                  <div class="p-2 bg-black/20 rounded space-y-1.5 mb-1.5">
-                      <div class="flex items-center space-x-2">
-                          <input type="text" bind:value={pattern.name} class="glass-input flex-grow !text-xs !bg-neutral-500/40" placeholder="Field Name (e.g., InvoiceID)"/>
-                          <Button variant="ghost" customClass="!px-1.5 !py-0.5 !text-red-400 hover:!text-red-300" on:click={() => stage.config.parameters.patterns = stage.config.parameters.patterns.filter(p => p.id !== pattern.id)}>X</Button>
-                      </div>
-                      <input type="text" bind:value={pattern.regex} class="glass-input w-full !text-xs !bg-neutral-500/40" placeholder="Regex Pattern (e.g., INV-\d+)"/>
-                      <!-- New Input for Capture Group Index -->
-                      <div>
-                        <label for={`pattern-group-index-${pattern.id}`} class="block text-xs font-medium text-gray-400 mb-0.5 mt-1">
-                          Capture Group Index
-                        </label>
-                        <input
-                          type="number"
-                          id={`pattern-group-index-${pattern.id}`}
-                          bind:value={pattern.captureGroupIndex}
-                          min="0"
-                          step="1"
-                          class="glass-input w-full !text-xs !bg-neutral-500/40"
-                          placeholder="e.g., 1 (0 for full match)"
-                        />
-                        <p class="text-sm font-light text-gray-500 dark:text-gray-400 mt-0.5">Default: 1 (first group). Use 0 for full match.</p>
-                      </div>
-                  </div>
-                {/each}
-                <Button variant="secondary" customClass="!text-xs !py-1" on:click={() => stage.config.parameters.patterns = [...stage.config.parameters.patterns, {id: Date.now().toString() + Math.random().toString(36).substring(2,9), name: '', regex: '', captureGroupIndex: 1}]}>Add Regex Pattern</Button>
-              </div>
-            {/if}
-
-            <!-- Parameters for SimpleTableExtraction -->
-            {#if stage.config?.strategy === 'SimpleTableExtraction'}
-              <div class="pl-3 border-l-2 border-neutral-700 space-y-2 py-2 text-xs">
-                  <label class="block font-medium text-gray-300 mb-0.5">Header Keywords (comma-separated):</label>
-                  <input type="text" bind:value={stage.config.parameters._headerKeywordsString}
-                         on:input={() => stage.config.parameters.headerKeywords = (stage.config.parameters._headerKeywordsString || '').split(',').map(s => s.trim()).filter(s => s)}
-                         class="glass-input w-full !text-xs !bg-neutral-500/40" placeholder="e.g., Item, Qty, Price"/>
-                  <label class="block font-medium text-gray-300 mt-1 mb-0.5">Stop Keywords (optional, comma-separated):</label>
-                  <input type="text" bind:value={stage.config.parameters._stopKeywordsString}
-                         on:input={() => stage.config.parameters.stopKeywords = (stage.config.parameters._stopKeywordsString || '').split(',').map(s => s.trim()).filter(s => s)}
-                         class="glass-input w-full !text-xs !bg-neutral-500/40" placeholder="e.g., Total, Subtotal"/>
-              </div>
-            {/if}
-          </div>
-        {/if}
-
-        <!-- Report Stage Configuration -->
-        {#if stage.type.toLowerCase() === 'report'}
-          <div class="form-group mt-3 pt-3 border-t border-neutral-700/50 space-y-3">
-            <div>
-              <label for={`stage-report-template-${stage.id}`} class="block text-xs font-medium text-gray-300 mb-1">
-                Report Markdown Template
-              </label>
-              <textarea
-                id={`stage-report-template-${stage.id}`}
-                bind:value={stage.config.template}
-                rows={8}
-                class="glass-input w-full !text-sm custom-scrollbar !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-                placeholder="Enter Markdown template. Use {{placeholder.path}} for data. e.g., {{document_name}}, {{ai_result.summary}}"
-              ></textarea>
-              <p class="text-sm font-light text-gray-400 dark:text-gray-500 mt-1">
-                Available placeholders depend on data from previous stages (e.g., `ai_result`, `parse_result`) and job metadata (`document_name`, `job_id`).
-              </p>
-            </div>
-
-            <div>
-              <label for={`stage-report-summary-fields-${stage.id}`} class="block text-xs font-medium text-gray-300 mb-1">
-                JSON Summary Fields (Optional, comma-separated JSONPaths)
-              </label>
-              <input
-                type="text"
-                id={`stage-report-summary-fields-${stage.id}`}
-                bind:value={stage.config._summaryFieldsString}
-                on:input={() => {
-                    if (!stage.config) stage.config = { template: '', summaryFields: [], _summaryFieldsString: '' }; // Ensure config exists
-                    stage.config.summaryFields = (stage.config._summaryFieldsString || '')
-                                                    .split(',')
-                                                    .map(s => s.trim())
-                                                    .filter(s => s);
-                    pipeline.stages = [...pipeline.stages]; // Force reactivity
-                }}
-                class="glass-input w-full !text-sm !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100"
-                placeholder="e.g., previous_stage_output.field.name, ai_result.sentiment"
-              />
-              <p class="text-sm font-light text-gray-400 dark:text-gray-500 mt-1">
-                Enter JSONPath expressions (e.g., `previous_stage_output.field.name`) relative to the data available to the report stage. Each path creates a key in the summary JSON.
-              </p>
-            </div>
-          </div>
-        {/if}
-      </div>
-    {/each}
+    <div class="space-y-3">
+      <StageList bind:stages={pipeline.stages} {availablePromptTemplates} {isLoadingOrgSettings} />
+    </div>
     <div class="flex gap-2 mt-3 p-3 border-t border-neutral-700/50">
       <input class="glass-input flex-1 !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100" bind:value={newStageType} placeholder="New Stage Type" />
       <input class="glass-input flex-1 !bg-neutral-600/50 !border-neutral-500/70 !text-gray-100" bind:value={newCommand} placeholder="Command (optional)" />
